refactor(#126): 서버 컴포넌트로 변환
